+++ AmpA.h
@@ -6,8 +6,8 @@ class AmpA
 	public:
 		//resistance and capacity value in Ohm and Fahrad
 		AmpA(
-		double r1 = 1000.0,
-		double c1 = 100.0e-12
+		long double r1 = 1000.0,
+		long double c1 = 100.0e-12
 		):
 			R1(r1), C1(c1)
 		{
@@ -15,17 +15,17 @@ class AmpA
 		}
 		~AmpA() {}
 
-		double amp  (double, double);
-		double ampID(double, double);
+		long double amp  (long double, long double);
+		long double ampID(long double, long double);
 
 		//RESET function set all parameters to the initial value
 		void RESET();
 
 	private:
-		double R1;
-		double C1;
+		long double R1;
+		long double C1;
 		//previous output of the amplifier
-		double V_old;
+		long double V_old;
 };
 
 #endif
