+++ AmpB.h
@@ -5,12 +5,12 @@ class AmpB
 {
 	public:
 		AmpB(
-		double r1 = 333.0,
-		double r2 = 6.7e3,
-		double r = 0.01,
-		double rout = 4.3e8,
-		double cout = 3.0e-15,
-		double g = 600.0e3
+		long double r1 = 333.0,
+		long double r2 = 6.7e3,
+		long double r = 0.01,
+		long double rout = 4.3e8,
+		long double cout = 3.0e-15,
+		long double g = 600.0e3
 		):
 		R1(r1), R2(r2), R(r), Rout(rout), Cout(cout), G(g)
 		{
@@ -18,22 +18,22 @@ class AmpB
 		}
 
 		//amplifier stage and IDeal implifier stage
-		double amp (double, double);
+		long double amp (long double, long double);
 		//ideal
-		double ampID(double, double);
+		long double ampID(long double, long double);
 
 		//RESET function set all parameters to the initial value
                 void RESET();
 
 
 	private:
-		double	R1;
-		double	R2;
-		double	Rout;
-		double	R;
-		double	Cout;
-		double	G;
-		double	Vout_old;
+		long double	R1;
+		long double	R2;
+		long double	Rout;
+		long double	R;
+		long double	Cout;
+		long double	G;
+		long double	Vout_old;
 };
 
 #endif
