+++ AmpF.cpp
@@ -1,14 +1,14 @@
 #include"AmpF.h"
 
-void AmpF::step1(double Vin, double dt)
+void AmpF::step1(long double Vin, long double dt)
 {
-double        a = (1.0/R+1.0/Rout)/Cout;
-double        S = -1.0/Cout*(1.0/R1+G/R);
-double    theta = 1.0/(R1*Cout);
-double    alpha = 1.0/R1+R2/R1/R3;
-double    gamma = C1*(R2/R1);
-double    beta1 = 1.0/R1+(R1+R2)/R1/R3;
-double    delta = C1/R1*(R1+R2);
+long double        a = (1.0/R+1.0/Rout)/Cout;
+long double        S = -1.0/Cout*(1.0/R1+G/R);
+long double    theta = 1.0/(R1*Cout);
+long double    alpha = 1.0/R1+R2/R1/R3;
+long double    gamma = C1*(R2/R1);
+long double    beta1 = 1.0/R1+(R1+R2)/R1/R3;
+long double    delta = C1/R1*(R1+R2);
 
 V1 = (V1_old*(1.0-beta1/(C1*S)-delta/(C1*dt*S))
 -Vin*(delta*theta/(C1*S)+dt*beta1*theta/(C1*S)+gamma/C1+dt*(alpha/C1))
@@ -22,14 +22,14 @@ V1 = (V1_old*(1.0-beta1/(C1*S)-delta/(C1*dt*S))
 }
 
 
-double AmpF::amp(double Vin, double dt)
+long double AmpF::amp(long double Vin, long double dt)
 {
 //step1 is called first to update V1 value
 step1(Vin,dt);
 
-double        a = (1.0/R+1.0/Rout)/Cout;
-double        S = -1.0/Cout*(1.0/R1+G/R);
-double    theta = 1.0/(R1*Cout);
+long double        a = (1.0/R+1.0/Rout)/Cout;
+long double        S = -1.0/Cout*(1.0/R1+G/R);
+long double    theta = 1.0/(R1*Cout);
 
 Vout = V1*(1.0L+dt*a)/(dt*S)-V1_old/(dt*S)-(theta/S)*Vin;
 
@@ -41,7 +41,7 @@ return Vout;
 }
 
 
-double AmpF::ampID(double Vin, double dt)
+long double AmpF::ampID(long double Vin, long double dt)
 {
 	Vout = (Vout_old - (R2+R3)/R1/R3*dt*Vin - R2/R1*(Vin-Vin_old))
 		/(1.0+dt/C1/R3);
