+++ hpf.cpp
@@ -1,8 +1,8 @@
 #include"hpf.h"
 
-double hpf::filter(double Iin, double dt)
+long double hpf::filter(long double Iin, long double dt)
 {
-	double Vout;
+	long double Vout;
 	Vout = (V_old + R*(Iin - I_old))/(1.0L + dt*R/L )/R; 
 	V_old = Vout;
 	I_old = Iin;
