+++ laser.cpp
@@ -12,16 +12,16 @@ void laser::Init()
 {
 	FREQ = C/LAMBDA;
 	FREQ0 = FREQ;
-        OMEGA = 2.0*PI*FREQ;
-        K = 2.0*PI/LAMBDA;
+        OMEGA = 2.0L*PI*FREQ;
+        K = 2.0L*PI/LAMBDA;
 }
 
-electric_field laser::GetField(double d, long double t)
+electric_field laser::GetField(long double d, long double t)
 {
-	double phi, In;
+	long double phi, In;
 	electric_field ef;
 	//phase of the emitted field when it reaches x at time t
-	phi = K*d - OMEGA*(t - d/(2.0*C)) + oscillator(t - d/(2.0*C));
+	phi = K*d - OMEGA*(t - d/(2.0L*C)) + oscillator(t - d/(2.0L*C));
 	//intensity of the field
 	In = INTENSITY;
 	ef.SetA(In);
@@ -29,14 +29,14 @@ electric_field laser::GetField(double d, long double t)
 	return ef;
 }
 
-double laser::oscillator(double t)
+long double laser::oscillator(long double t)
 {
-double dphi = beta*sin(omegaM*t);
+long double dphi = beta*sin(omegaM*t);
 return dphi;
 }
 
 
-void laser::ErrSig(double err)
+void laser::ErrSig(long double err)
 {
 	SetFreq(FREQ + err);
 }
